feat(scripts): update add_specs.sh script
Since most teams inside MagaLu only want to update their own spec file,
the add_specs.sh was changed to receive a new openapi.yaml file, execute
all transformations and add in the correct openapi folder.

The spec transformations were also refactored into different files,
making it easier to create new transformers, and allowing the chaining
of those operations.
