+++ scripts/transformers/spec_types.py
@@ -1,0 +1,8 @@
+from typing import Any, Dict, Protocol
+
+OAPISchema = Dict[str, Any]
+
+
+class SpecTranformer(Protocol):
+    def transform(self, spec: OAPISchema) -> OAPISchema:
+        pass
